Version 4.6 (2017-09-25) Bugs fixed:
Don’t mark a drive order as finished if the transport order it belongs to was withdrawn.

Properly update the vehicles' states in the kernel control center’s vehicle list.

When creating locations, properly attach links to the respective points, too.

When renaming a point in the plant overview, properly update blocks containing paths starting or ending at this point.

Avoid NPE when the transport order referenced in a Vehicle instance does not exist in the kernel any more.

New features:

Allow the kernel’s RMI port to be set via configuration.

Allow preferred parking positions and recharge locations to be set as properties on Vehicle instances.

In XML status channel messages, add a reference to a vehicle’s transport order, and vice versa.

Allow the kernel’s order cleanup task to be adjusted via predicates that approve cleanup of transport orders and order sequences.

Other changes:

Deprecate VehicleCommAdapter.State. It’s not really used anywhere, and the enum elements are fuzzy/incomplete, anyway.

中文翻译(机翻警告)
版本4.6（2017-09-25）
错误修复：

如果驱动器订单属于其运输订单，则不要将其标记为已完成。

在内核控制中心的车辆清单中正确更新车辆的状态。

创建位置时，也请正确将链接附加到各个点。

在工厂概览中重命名点时，请正确更新包含在此点开始或结束的路径的块。

当Vehicle实例中引用的传输顺序在内核中不再存在时，请避免使用NPE。

新的功能：

允许通过配置设置内核的RMI端口。

允许将首选停车位置和充电位置设置为“车辆”实例的属性。

在XML状态频道消息中，添加对车辆运输顺序的引用，反之亦然。

允许通过批准清理运输订单和订单序列的谓词来调整内核的订单清理任务。

其他变化：

弃用VehicleCommAdapter.State。它并没有真正在任何地方使用，枚举元素仍然是模糊/不完整的。