Version 4.9 (2018-04-16) Bugs fixed:
Fix jumping mouse cursor when dragging/moving model elements in the Plant Overview in some cases.

New features:

Allow the kernel to work headless, i.e. without a GUI. Introduce a separate Kernel Control Center application that provides the same functionality and can be attached to the kernel as a client.

Provide a single-threaded executor for sequential processing of tasks in the kernel, which helps avoiding locking and visibility issues. Use this executor for most tasks, especially the ones manipulating kernel state, that were previously executed concurrently.

Introduce a web API (HTTP + JSON), intended to replace the proprietary TCP/IP host interface, which is now deprecated.

Introduce an API for pluggable model import and export implementations in the Plant Overview.

Other changes:

Split the Kernel interface into aspect-specific service interfaces.

Provide a (more) simple event API, including an event bus implementation as a replacement for the previously used MBassador and event hub.

Overhaul the default dispatcher implementation to improve maintainability and extensibility.

Allow suggestions for property values in the Plant Overview to depend on the key.

Improve API and deprecate classes and methods in lots of places.

Improve default formatting of log output for better readability.

中文翻译(机翻警告)
版本4.9（2018-04-16）
错误修复：

在某些情况下，当在Plant Overview中拖动/移动模型元素时，修复了鼠标跳动的问题。

新的功能：

让内核无头工作，即没有GUI。介绍一个单独的Kernel Control Center应用程序，该应用程序提供相同的功能，并且可以作为客户端附加到内核。

提供一个单线程执行程序，用于内核中任务的顺序处理，这有助于避免锁定和可见性问题。将此执行程序用于大多数任务，尤其是先前并发执行的操纵内核状态的任务。

引入一个Web API（HTTP + JSON），用于替换专有的TCP / IP主机接口，该接口现已不推荐使用。

在工厂概述中引入用于可插入模型导入和导出实现的API。

其他变化：

将内核接口拆分为特定于方面的服务接口。

提供（更多）简单的事件API，包括事件总线实现，以替代以前使用的MBassador和事件中心。

全面检查默认调度程序实现，以提高可维护性和可扩展性。

允许“工厂概览”中有关属性值的建议取决于该键。

改进API，并在许多地方弃用类和方法。

改进日志输出的默认格式，以提高可读性。