Version 4.10 (2018-08-07) New features:
Introduce an explicit integration level property for vehicles that expresses to what degree a vehicle should be integrated into the system. (Setting the integration level to TO_BE_UTILIZED replaces the manual dispatching that was previously used to integrate a vehicle.)

Allow recomputing of a vehicle’s route after finishing a drive order or on topology changes.

Allow vehicle themes to define not only the graphics used, but also the content and style of vehicle labels in the Plant Overview.

Enable the web API to optionally use HTTPS.

Allow an optional set of properties for meta information to be stored in a model, and use it to store the model file’s last-modified time stamp in it.

Bugs fixed:

Prevent moving of model elements in the Plant Overview when in mode OPERATING.

Prevent creation of groups in the Plant Overview when in mode OPERATING.

Properly handle renaming of paths and path names that do not follow the default naming pattern in the Plant Overview.

Multiple minor fixes for the integration project generator.

Other changes:

When using the Plant Overview or Kernel Control Center with SSL-encrypted RMI, verification of the server certificate is now mandatory.

Adjust the default docking frames layout in the Plant Overview for mode OPERATING a bit to make better use of wide-screen displays.

Include web API documentation generated by Swagger in the distribution.
中文翻译(机翻警告)
版本4.10（2018-08-07）
新的功能：

为车辆引入一个明确的集成级别属性，该属性表示车辆应集成到系统中的程度。 （将集成级别设置为TO_BE_UTILIZED将替换以前用于集成车辆的手动调度。）

在完成驾驶命令或拓扑更改后，允许重新计算车辆的路线。

允许车辆主题不仅定义所使用的图形，还可以定义“工厂概览”中车辆标签的内容和样式。

启用Web API以选择使用HTTPS。

允许将元信息的一组可选属性存储在模型中，并使用它在其中存储模型文件的最后修改的时间戳。

错误修复：

在操作模式下，防止在“工厂概览”中移动模型元素。

在操作模式下，防止在工厂概览中创建组。

正确处理不遵循“工厂概览”中默认命名模式的路径和路径名称的重命名。

集成项目生成器的多个次要修复程序。

其他变化：

当使用带有SSL加密的RMI的Plant Overview或Kernel Control Center时，现在必须对服务器证书进行验证。

在“设备概述”中将模式的默认对接框架布局调整为“操作”模式，以更好地利用宽屏显示。

在发行版中包含Swagger生成的Web API文档。