Version 4.11 (2018-12-04) New features:
Introduce a type property for blocks. A block’s type now determines the rules for entering it:

Single vehicle only: The resources aggregated in this block can only be used by a single vehicle at the same time.

Same direction only: The resources aggregated in this block can be used by multiple vehicles, but only if they enter the block in the same direction.

Bugs fixed:

Properly set a point’s layout coordinates when it is placed exactly on an axis in the Plant Overview.

Properly select the correct/clicked-on tree entry in the Plant Overview’s blocks tree view when the same element is a member of more than one block.

Prevent the Kernel application from freezing when loading some larger plant models.

Other changes:

Require the user to confirm immediate withdrawals of transport orders in the plant overview, as they have some implications that may lead to collisions or deadlocks in certain situations.

Improve input validation of unit-based properties for plant model elements.

Remove the Kernel Control Center’s function to reset the position of a vehicle. Users should now set the vehicle’s integration level to TO_BE_IGNORED, instead.

Allow the loopback driver to be disabled completely.

Minor improvements to the configuration interface API.

Mark all AdapterCommand implementations in the base API as deprecated. These commands' functionality is specific to the respective communication adapter and should be implemented and used there.
中文翻译(机翻警告)
版本4.11（2018-12-04）
新的功能：

介绍块的类型属性。现在，方块的类型决定了输入方块的规则：

仅单车：此块中聚合的资源只能同时由单车使用。

仅限同一方向：此块中聚集的资源可由多辆车使用，但前提是它们以相同方向进入该块。

错误修复：

将点精确放置在“工厂概览”中的轴上时，请正确设置该点的布局坐标。

当同一元素是多个块的成员时，在“工厂概览”的块树视图中正确选择正确/单击的树条目。

加载一些较大的工厂模型时，防止内核应用程序冻结。

其他变化：

要求用户在工厂概览中确认立即撤回运输订单，因为它们可能会在某些情况下导致碰撞或死锁。

改进工厂模型元素的基于单位的属性的输入验证。

删除内核控制中心的功能以重置车辆的位置。用户现在应该将车辆的集成度设置为TO_BE_IGNORED。

允许完全禁用回送驱动程序。

对配置接口API的较小改进。

将基本API中的所有AdapterCommand实现标记为已弃用。这些命令的功能特定于相应的通信适配器，应在此处实现和使用。