Version 4.13 (2019-02-18) New features:
Introduce an event history for transport orders that can be filled with arbitrary event data.

Introduce "*" as a wildcard in a vehicle’s processable categories to allow processing of transport orders in any category.

The Plant Overview’s vehicle panel now also shows the current destination of each vehicle.

Bugs fixed:

With the defaultdispatcher.rerouteTrigger configuration entry set to DRIVE_ORDER_FINISHED, ensure that the rerouting is only applied to the vehicle that has actually finished a drive order.

For vehicles selected in the Plant Overview, re-allow changing their integration levels via the context menu to either "to be utilized" or "to be respected" if any of them is currently processing a transport order, too.

Other changes:

Remove the included integration project generator and document usage of the example integration project, instead.

Update the web API specification to OpenAPI 3.

Update Gradle to 4.10.3.

Update Checkstyle to 8.16.

Update JUnit to 5.3.2.

Update Guice to 4.2.2.

中文翻译(机翻警告)
版本4.13（2019-02-18）
新的功能：

介绍运输历史记录的事件历史记录，该历史记录可以填充任意事件数据。

在车辆的可处理类别中引入“ *”作为通配符，以允许处理任何类别的运输订单。

现在，“工厂概览”的车辆面板还显示了每辆车的当前目的地。

错误修复：

在将defaultdispatcher.rerouteTrigger配置条目设置为DRIVE_ORDER_FINISHED的情况下，确保重新路由仅应用于实际完成驾驶命令的车辆。

对于在“工厂概览”中选择的车辆，如果其中任何一个当前也在处理运输订单，则可以通过上下文菜单重新将其集成级别更改为“要使用”或“要遵守”。

其他变化：

而是删除包含的集成项目生成器，并代替示例集成项目的文档用法。

将Web API规范更新为OpenAPI 3。

将Gradle更新为4.10.3。

将Checkstyle更新为8.16。

将JUnit更新到5.3.2。

将Guice更新至4.2.2。