Version 4.16 (2019-10-22) New features:
Optionally have names for transport orders and order sequences generated by the kernel. Use ULIDs for these generated names by default, to have lexicographically sortable names.

Add a publishEvent() method to the KernelServicePortal interface that RMI-Clients can use to publish events on the Kernel application’s event bus.

Enable the Kernel Control Center application to set positions for all simulating vehicle drivers, not only the loopback driver.

Bugs fixed:

Paths that have the same start and end components are now displayed properly in the Plant Overview.

In the Plant Overview’s continuous load panel, transport order definitions can now be saved to and restored from XML files again. (Note that in the course of fixing this issue, the XML files' structure was improved. Since the feature had been broken for a while and is not part of a public API, backwards compatibility was not maintained for this. As a result, transport order definition files from old versions of openTCS cannot be restored.)

Make using the "try it out" buttons in the OpenAPI documentation possible by setting CORS headers in the web API’s responses.

中文翻译(机翻警告)
版本4.16（2019-10-22）
新的功能：

（可选）具有内核生成的运输订单和订单序列的名称。默认情况下，将ULID用于这些生成的名称，以具有按词典顺序可排序的名称。

向KernelServicePortal接口添加publishEvent（）方法，RMI客户端可使用该方法在Kernel应用程序的事件总线上发布事件。

启用内核控制中心应用程序以设置所有模拟车辆驾驶员（而不只是环回驾驶员）的位置。

错误修复：

现在，在工厂概览中可以正确显示具有相同起始和结束成分的路径。

现在，在“工厂概览”的“连续装载”面板中，可以再次将运输订单定义保存到XML文件或从XML文件恢复。 （请注意，在解决此问题的过程中，改进了XML文件的结构。由于该功能已被破坏了一段时间，并且不是公共API的一部分，因此未维护向后兼容性。因此，传输旧版openTCS的订单定义文件无法恢复。）

通过在Web API的响应中设置CORS标头，可以使用OpenAPI文档中的“试用”按钮。